In clasa Board am corectat getteri si move constructori + alte mici modificari pentru explozii

+++ Eter/Eter/bmode.cpp
@@ -18,7 +18,7 @@ BMode::BMode(Game* game):
 
 void BMode::applyModeRules()
 {
-    m_game->GetBoardRef().SetDimMax((uint8_t)4);
+    m_game->GetBoardRef().SetDimMax(4);
     generateMage();
     AMode::applyModeRules();
 }
